+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/balances/AndroidBalances.kt
@@ -3,16 +3,22 @@
 package co.electriccoin.zcash.ui.screen.balances
 
 import androidx.activity.ComponentActivity
+import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.screen.balances.view.Balances
 
-@Suppress("UNUSED_PARAMETER")
 @Composable
 internal fun WrapBalances(
     activity: ComponentActivity,
     goSettings: () -> Unit,
 ) {
+    val walletViewModel by activity.viewModels<WalletViewModel>()
+    val walletSnapshot = walletViewModel.walletSnapshot.collectAsStateWithLifecycle().value
+
     Balances(
+        walletSnapshot = walletSnapshot,
         onSettings = goSettings,
     )
 }