+++ CHANGELOG.md
@@ -9,12 +9,17 @@ directly impact users rather than highlighting other key architectural updates.*
 
 ## [Unreleased]
 
+## [0.2.0 (540)] - 2024-01-27
+
 ### Added
 - The current balance UI on top of the Account screen has been reworked. It now displays the currently available 
   balance as well.
 - The same current balance UI was incorporated into the Send and Balances screens. 
 - The Send Error screen now contains a simple text with the reason for failure. The Send screen UI refactoring is 
   still in progress.
+
+### Fixed
+- Properly clearing focus from the Send text fields when moved to another screen
 
 ## [0.2.0 (530)] - 2024-01-16